+++ KenHRApp.Domain/Entities/AttendanceSummaryResult.cs
@@ -21,7 +21,7 @@ namespace KenHRApp.Domain.Entities
         public int TotalEarlyOut { get; set; }
         public decimal TotalDeficitHour { get; set; }
         public decimal TotalWorkHour { get; set; }
-        public decimal TotalDaysWorked { get; set; }
+        public int TotalDaysWorked { get; set; }
         public decimal AverageWorkHour { get; set; }
         public decimal TotalLeaveBalance { get; set; }
         public decimal TotalSLBalance { get; set; }
